client-sideに合わせてserver-sideを調整

+++ app/controllers/api/admin_controller.rb
@@ -9,21 +9,26 @@ class Api::AdminController < Api::ApplicationController
   end
 
   def create
-    user = User.create!(user_params)
+    User.create!(user_params)
     render json: { message: '新規ユーザー登録完了' }, status: :ok
   end
 
   def update
-    user = User.update!(user_params)
+    User.find(params[:id]).update!(user_params)
     render json: { message: 'ユーザー更新完了' }, status: :ok
   end
 
   def destroy
-    user = User.find(params[:id])
-    tasks = user.tasks
-    user.destroy!
-    tasks.destroy_all
-    render json: { message: 'ユーザー、及び関連するタスク削除完了' }, status: :ok
+    current_user = User.find_by(token: session[:token])
+    if current_user.id == params[:id]
+      render json: { message: '自分で自分のアカウントは削除できません' }, status: :unprocessable_entity
+    else
+      user = User.find(params[:id])
+      tasks = user.tasks
+      user.destroy!
+      tasks.destroy_all
+      render json: { message: 'ユーザー、及び関連するタスク削除完了' }, status: :ok
+    end
   end
 
   private
@@ -33,6 +38,6 @@ class Api::AdminController < Api::ApplicationController
   end
 
   def user_params
-    params.require(:user).permit(:name, :email, :password_digest)
+    params.require(:user).permit(:name, :email, :permission, :admin)
   end
 end
